requirements.txt
networkx==3.3
python-dotenv==1.1.0
Requests==2.32.3
openai==1.57.4
google-generativeai==0.6.0
prettytable==3.16.0
docker==7.1.0
pandas==2.3.0
litellm
legacy-cgi==2.6.2